refactor(api): moves all refactories into same folder (BOAS-52)
