+++ apps/api/src/server/applications/case-officer/case-officer.controller.js
@@ -1,8 +1,8 @@
-import * as applicationRepository from '../repositories/application.repository.js';
-import { mapApplicationsWithSectorAndSubSector } from './case-officer.mapper.js'
+import * as applicationRepository from '../../repositories/application.repository.js';
+import { mapApplicationsWithSectorAndSubSector } from './case-officer.mapper.js';
 
-export const getApplications = async(_request, response) => {
-    const applications = await applicationRepository.getByStatus('open');
+export const getApplications = async (_request, response) => {
+	const applications = await applicationRepository.getByStatus('open');
 
-    return response.send(mapApplicationsWithSectorAndSubSector(applications));
-}
+	return response.send(mapApplicationsWithSectorAndSubSector(applications));
+};
