+++ apps/api/src/server/repositories/application.repository.js
@@ -1,4 +1,4 @@
-import { databaseConnector } from '../../utils/database-connector.js';
+import { databaseConnector } from '../utils/database-connector.js';
 
 /**
  * @param {string} status
